# openapi-v2.json
{"openapi":"3.0.3","info":{"title":"Airports API","description":"\n![Logo](https://raw.githubusercontent.com/traefik-workshops/hub-tutorials/master/assets/traefik-airelines-logo.png)\n\n## Overview\n\nThis API can be used to get vital information on major international airports.\n\n- Name of the airport\n- City or the airport\n\n## Usage\n\nRetrieve a list of all airports by sending a `GET` request to `/airports`.\n\n```shell\ncurl --request GET --url https://api.traefik-airlines/v1/airports/airports --header 'Accept: application/json'\n```\n\nIf you want to request information on a certain airport use the airport `ID`.\n\n```shell\ncurl --request GET --url https://api.traefik-airlines/v1/airports/airports/{ID} --header 'Accept: application/json'\n```\n\n### Response codes\n\nIf the status returned is in the `200` range, it indicates that the request was fulfilled successfully and that no error was encountered.\n\nReturn codes in the `400` range indicate that there was an issue with the request that was sent.\n\n### Versioning\n\nThis API uses semantic versioning to ensure that your client doesn't break.  \nThe version is declared in the URL so that you can lock to a specific one by prefix the URL.\n\nWhen the version declared in the URL is not supported, you will receive a 400 response.\n\n\nThis API helps you do awesome travel stuff. 🚀\n\n\nList some airports:\n\n- Amsterdam\n- Lyon\n- Reykjavík\n","version":"2.0.0","x-logo":{"url":""}},"paths":{"/":{"get":{"tags":["About"],"summary":"Welcome","description":"Welcome message","operationId":"Index__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Index  Get"}}}}}}},"/airports/":{"get":{"tags":["Airports"],"summary":"List all airports","description":"Overview of all airports supported by this API.","operationId":"read_item_airports__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/airports/{airport_id}":{"get":{"tags":["Airports"],"summary":"Receive airport by ID","description":"Fetch a single airport by ID","operationId":"fetch_airport_airports__airport_id__get","parameters":[{"name":"airport_id","in":"path","required":true,"schema":{"type":"integer","title":"Airport Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Fetch Airport Airports  Airport Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["Utils"],"summary":"Check health","description":"Check service status","operationId":"perform_healthcheck_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}